Change settings route from `/admin/settings` to `/settings`
