+++ app/components/nav_bar/nav_bar.html.erb
@@ -17,7 +17,7 @@
         <% end %>
       </li>
       <li class="NavBar-item">
-        <%= link_to admin_settings_path do %>
+        <%= link_to settings_path do %>
           <%= t 'setting.setting', count: :many %>
         <% end %>
       </li>
