+++ app/components/settings_form/settings_form.html.erb
@@ -1,4 +1,4 @@
-<%= form_with(url: admin_settings_path, local: true, class: 'SettingsForm' ) do |f|%>
+<%= form_with(url: settings_path, local: true, class: 'SettingsForm' ) do |f|%>
   <%= c 'stack', space: :xlarge do %>
 
     <%= c 'stack' do %>
