+++ config/routes.rb
@@ -10,6 +10,9 @@ Rails.application.routes.draw do
   post '/onboarding', to: 'onboarding#create'
   get '/onboarding/success', to: 'onboarding#success'
 
+  get '/settings', to: 'settings#index'
+  post '/settings', to: 'settings#update'
+
   telegram_webhook Telegram::WebhookController
 
   resources :requests, only: %i[index show new create] do
@@ -34,9 +37,5 @@ Rails.application.routes.draw do
     member do
       post 'highlight', format: /json/
     end
-  end
-
-  namespace :admin do
-    resources :settings, only: %i[index create]
   end
 end
